+++ ToolGood.ReadyGo3/Gadget/TableManager/ColumnInfo.cs
@@ -23,7 +23,7 @@ namespace ToolGood.ReadyGo3.Gadget.TableManager
         internal static ColumnInfo FromProperty(PropertyInfo pi)
         {
             if (pi.CanRead == false || pi.CanWrite == false) return null;
-            if (IsAllowType(pi.PropertyType) == false) return null;
+            if (Types.IsAllowType(pi.PropertyType) == false) return null;
             var a = pi.GetCustomAttributes(typeof(IgnoreAttribute), true);
             if (a.Length > 0) return null;
             a = pi.GetCustomAttributes(typeof(ResultColumnAttribute), true);
@@ -54,84 +54,12 @@ namespace ToolGood.ReadyGo3.Gadget.TableManager
                 if (pi.PropertyType == typeof(string) || pi.PropertyType == typeof(AnsiString)) {
                     ci.Required = false;
                 } else {
-                    ci.Required = IsNullType(ci.PropertyType) == false;
+                    ci.Required = Types.IsNullType(ci.PropertyType) == false;
                 }
             }
-            ci.PropertyType = GetBaseType(ci.PropertyType);
-
+            ci.PropertyType = Types.GetBaseType(ci.PropertyType);
             return ci;
         }
-        private static bool IsAllowType(Type type)
-        {
-            if (type == null) return false;
-            if (type.IsEnum) return true;
-            if (type == typeof(byte[])) return true;
-            if (type == typeof(sbyte[])) return true;
-            if (type.FullName == "Microsoft.SqlServer.Types.SqlGeography") return true;
-            if (type.FullName == "Microsoft.SqlServer.Types.SqlGeometry") return true;
 
-            if (type.IsGenericType) {
-                if (type.GetGenericTypeDefinition().Equals(typeof(Nullable<>))) {
-                    type = type.GetGenericArguments()[0];
-                } else {
-                    return false;
-                }
-            }
-
-            if (type == typeof(Guid)) return true;
-            if (type == typeof(AnsiString)) return true;
-            if (type == typeof(TimeSpan)) return true;
-            if (type == typeof(DateTimeOffset)) return true;
-
-            var tc = Type.GetTypeCode(type);
-            switch (tc) {
-                case TypeCode.Boolean:
-                case TypeCode.Byte:
-                case TypeCode.Char:
-                case TypeCode.DateTime:
-                case TypeCode.Decimal:
-                case TypeCode.Double:
-                case TypeCode.Int16:
-                case TypeCode.Int32:
-                case TypeCode.Int64:
-                case TypeCode.SByte:
-                case TypeCode.Single:
-                case TypeCode.String:
-                case TypeCode.UInt16:
-                case TypeCode.UInt32:
-                case TypeCode.UInt64:
-                    return true;
-
-                case TypeCode.Object:
-                default:
-                    break;
-            }
-            return false;
-        }
-
-        private static bool IsNullType(Type type)
-        {
-            if (type == null) return true;
-            if (type.IsEnum) return false;
-
-            if (type.IsGenericType) {
-                if (type.GetGenericTypeDefinition().Equals(typeof(Nullable<>))) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        private static Type GetBaseType(Type type)
-        {
-            if (type.IsEnum) return type;
-
-            if (type.IsGenericType) {
-                if (type.GetGenericTypeDefinition().Equals(typeof(Nullable<>))) {
-                    type = type.GetGenericArguments()[0];
-                }
-            }
-            return type;
-        }
     }
 }
