+++ ToolGood.ReadyGo3/PetaPoco/Core/ColumnInfo.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Reflection;
 using ToolGood.ReadyGo3.Attributes;
+using ToolGood.ReadyGo3.Gadget;
 using ToolGood.ReadyGo3.PetaPoco.Internal;
 
 namespace ToolGood.ReadyGo3.PetaPoco
@@ -49,7 +50,7 @@ namespace ToolGood.ReadyGo3.PetaPoco
         internal static ColumnInfo FromProperty(PropertyInfo pi)
         {
             if (pi.CanRead == false || pi.CanWrite == false) return null;
-            if (IsAllowType(pi.PropertyType) == false) return null;
+            if (Types.IsAllowType(pi.PropertyType) == false) return null;
 
             var explicitColumns = pi.DeclaringType.GetCustomAttributes(typeof(ExplicitColumnsAttribute), true).Length > 0;
             // Check for [Column]/[Ignore] Attributes
@@ -85,57 +86,6 @@ namespace ToolGood.ReadyGo3.PetaPoco
             }
             return ci;
         }
-
-        internal static Cache<Type, bool> _IsAllowType = new Cache<Type, bool>();
-        internal static bool IsAllowType(Type type)
-        {
-            return _IsAllowType.Get(type, () => {
-                if (type == null) return false;
-                if (type.IsEnum) return true;
-                if (type == typeof(byte[])) return true;
-                if (type == typeof(sbyte[])) return true;
-                if (type.FullName == "Microsoft.SqlServer.Types.SqlGeography") return true;
-                if (type.FullName == "Microsoft.SqlServer.Types.SqlGeometry") return true;
-
-                if (type.IsGenericType) {
-                    if (type.GetGenericTypeDefinition().Equals(typeof(Nullable<>))) {
-                        type = type.GetGenericArguments()[0];
-                    } else {
-                        return false;
-                    }
-                }
-
-                if (type == typeof(Guid)) return true;
-                if (type == typeof(AnsiString)) return true;
-                if (type == typeof(TimeSpan)) return true;
-                if (type == typeof(DateTimeOffset)) return true;
-
-                var tc = Type.GetTypeCode(type);
-                switch (tc) {
-                    case TypeCode.Boolean:
-                    case TypeCode.Byte:
-                    case TypeCode.Char:
-                    case TypeCode.DateTime:
-                    case TypeCode.Decimal:
-                    case TypeCode.Double:
-                    case TypeCode.Int16:
-                    case TypeCode.Int32:
-                    case TypeCode.Int64:
-                    case TypeCode.SByte:
-                    case TypeCode.Single:
-                    case TypeCode.String:
-                    case TypeCode.UInt16:
-                    case TypeCode.UInt32:
-                    case TypeCode.UInt64:
-                        return true;
-
-                    case TypeCode.Object:
-                    default:
-                        break;
-                }
-                return false;
-            });
-
-        }
+ 
     }
 }
